+++ src/main/java/com/rudkids/rudkids/domain/product/service/strategy/productStatus/ChangeProductStatusStrategy.java
@@ -1,4 +1,4 @@
-package com.rudkids.rudkids.domain.product.service.strategy;
+package com.rudkids.rudkids.domain.product.service.strategy.productStatus;
 
 import com.rudkids.rudkids.domain.product.domain.Product;
 import com.rudkids.rudkids.domain.product.domain.ProductStatus;
