+++ src/main/java/com/rudkids/rudkids/domain/product/service/strategy/productStatus/OpenStrategy.java
@@ -1,18 +1,18 @@
-package com.rudkids.rudkids.domain.product.service.strategy;
+package com.rudkids.rudkids.domain.product.service.strategy.productStatus;
 
 import com.rudkids.rudkids.domain.product.domain.Product;
 import com.rudkids.rudkids.domain.product.domain.ProductStatus;
 import org.springframework.stereotype.Component;
 
 @Component
-public class CloseStrategy implements ChangeProductStatusStrategy {
+public class OpenStrategy implements ChangeProductStatusStrategy {
     @Override
     public boolean isProductStatus(ProductStatus productStatus) {
-        return ProductStatus.CLOSED == productStatus;
+        return ProductStatus.OPEN == productStatus;
     }
 
     @Override
     public void changeStatus(Product product) {
-        product.close();
+        product.open();
     }
 }
